refactor: 레시피 상세 페이지 컴포넌트 개선 (#12)
* refactor: react-query 로직 삭제

* refactor: 수정 및 삭제 버튼 헤더로 통합

* refactor: 레시피 반응 컴포넌트 props 개선
